requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
databases==0.3.2
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
numpy==1.18.5
pandas==1.0.4
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
whitenoise==5.1.0
xlwt==1.3.0